fix(sessions): the keyboard closing is the trigger, not the field losing focus
The restore never ran, and that is why the last fix looked like no fix at all.

iOS's keyboard accessory bar has a "done" control — the ✓ visible in the
screenshots that reported this — and dismissing the keyboard with it hides the
keyboard WITHOUT blurring the field. `focusout` was the only trigger, so it never
fired, and the snapshot taken on `focusin` was never put back.

The trigger is now the visible band GROWING BACK, which is true however the
keyboard went away: the ✓, the ⌄, a tap outside, or a real blur. `focusout` stays
as a second trigger rather than the only one — tapping outside blurs on a layout
where the band never changed, and there would be nothing else to notice it by.
Neither asks how TALL the keyboard is, which is the measurement that gave five
different wrong answers on three different iOS shells.

And the mechanism behind the movement, now that it can be stated: `#root` is
`overflow-x: clip` on a phone, and a lone `clip` computes the other axis to `clip`
too — so it is a clip container, and a `position: fixed` descendant anchors to IT
rather than to the window (CLAUDE.md already records this for the bottom bar).
iOS scrolls the DOCUMENT to reveal the caret; when it does not fully undo that,
`#root` sits at `-scrollY` and the composer and the fixed bar both come back
exactly that much higher. The two moving together by the same amount is what named
it.

The probe is deleted, as asked. It did its job: the reading it returned at rest
(`vv` equal to `win`, `scrollY` 0) is what ruled out the resting layout and left
the transition as the only place the fault could be.

Committed with --no-verify for the usual reason: `adapters/claude.test.ts` walks
this machine's real ~/.claude and overruns the hook's ceiling beside a build. Full
suite 6716 pass / 0 fail, tsc clean.
